one big update

+++ src/main/java/eu/matejkormuth/starving/main/Time.java
@@ -42,7 +42,7 @@ public class Time {
     }
 
     public static Time ofTicks(long ticks) {
-        return new Time(ticks / 20 * 1000);
+        return new Time((long) (ticks / 20f * 1000f));
     }
 
     public static Time ofSeconds(long seconds) {
@@ -58,10 +58,10 @@ public class Time {
     }
 
     public int toTicks() {
-        return (int) (this.milis / 1000 * 20);
+        return (int) (this.milis / 1000f * 20f);
     }
 
     public long toLongTicks() {
-        return this.milis / 1000 * 20;
+        return (long) (this.milis / 1000f * 20f);
     }
 }
